everything done except unit test cmake and example
