+++ includes/02-extract/csv_reader.hpp
@@ -1,3 +1,13 @@
+/*
+
+    「 ✦ USA QuickPath Algorithm by TEAM 8 ✦ 」
+
+༺☆༻____________☾✧ ✩ ✧☽____________༺☆༻
+
+    STEP: 2/6 - EXTRACT
+    The files in the "extract" directory are responsible for converting the CSV data to binary format for faster processing.
+
+*/
 // ‿︵‿୨ Linking ୧‿︵‿
 
     #ifndef CSV_READER_H
@@ -14,13 +24,13 @@
     struct Road {
         int landmarkA;
         int landmarkB;
-        int time;
+        double time;
 
         // Default constructor
         Road() : landmarkA(0), landmarkB(0), time(0) {}
 
         // Parameterized constructor
-        Road(int a, int b, int t) : landmarkA(a), landmarkB(b), time(t) {}
+        Road(int a, int b, double t) : landmarkA(a), landmarkB(b), time(t) {}
     };
 
     /**
