+++ src/02-extract/binary_writer.cpp
